requirements.txt
numpy
pandas
scikit-learn
joblib
opencv-python
mediapipe
torch
torchaudio
torchvision
pyttsx3
matplotlib
pyserial
